requirements.txt
bcrypt
blinker
certifi
cffi
click
Flask
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
Pillow
pycparser
six
Werkzeug
WTForms
email-validator